added package for proportion and JavaClassesProjectBK.java to get java classes to  create dataset

+++ src/principale/LinkageBK.java
@@ -44,7 +44,7 @@ public class LinkageBK {
 			return -1; 
 		 }
 		 
-		 return (countTickets/count) ;		
+		 return ( (double)countTickets /count) ;		
 		}//try
 				 
 	}//fine metodo
